clf/pyproject.toml
[project]
name = "clf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "clip",
    "colored>=2.3.1",
    "huggingface-hub>=0.35.1",
    "hydra-core>=1.3.2",
    "lightning[pytorch-extra]>=2.5.5",
    "numpy>=2.3.3",
    "onnxruntime-gpu>=1.23.0",
    "onnxsim>=0.4.36",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "polygraphy>=0.49.26",
    "sam-2",
    "scikit-learn>=1.7.2",
    "tensorboard>=2.20.0",
    "timm>=1.0.20",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "transformers>=4.56.2",
    "ultralytics>=8.3.203",
]

[dependency-groups]
dev = [
    "jupyterlab>=4.4.7",
    "matplotlib>=3.10.6",
    "mdformat>=0.7.22",
    "pyclean>=3.1.0",
    "pytest>=8.4.2",
    "ruff>=0.13.2",
    "yamlfix>=1.18.0",
]

[tool.uv.sources]
clip = { git = "https://github.com/openai/CLIP.git" }
sam-2 = { git = "https://github.com/facebookresearch/sam2.git" }

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]
line-length = 120
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
extend-select = ["I"]
ignore = []
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"
